requirements.txt
flask
xlrd==1.2.0
xlsxwriter
mysql-connector-python
